event-hubs/requirements.txt
azure-eventhub==5.10.1
python-dotenv==0.21.0